requirements.txt
flask==0.12.2
requests==2.18.4
mysql-connector-python == 8.0.13